feature: build tenant management web api

+++ services/datntdev.Microservices.Srv.Admin.Web.App/MultiTenancy/TenantManager.cs
@@ -1,4 +1,5 @@
-﻿using datntdev.Microservices.Common.Web.App.Application;
+﻿using datntdev.Microservices.Common;
+using datntdev.Microservices.Common.Web.App.Application;
 using datntdev.Microservices.Common.Web.App.Exceptions;
 using datntdev.Microservices.Srv.Admin.Web.App.MultiTenancy.Models;
 using Microsoft.Extensions.DependencyInjection;
@@ -9,8 +10,11 @@ namespace datntdev.Microservices.Srv.Admin.Web.App.MultiTenancy
     {
         private readonly SrvAdminDbContext _dbContext = services.GetRequiredService<SrvAdminDbContext>();
 
+        public IQueryable<AppTenantEntity> GetQueryable() => _dbContext.AppTenants.AsQueryable();
+
         public override async Task<AppTenantEntity> CreateEntityAsync(AppTenantEntity entity)
         {
+            entity.Id = (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds();
             var createdEntity = await _dbContext.AppTenants.AddAsync(entity);
             await _dbContext.SaveChangesAsync();
             return createdEntity.Entity;
@@ -18,8 +22,12 @@ namespace datntdev.Microservices.Srv.Admin.Web.App.MultiTenancy
 
         public override async Task DeleteEntityAsync(int id)
         {
+            if (Constants.MultiTenancy.DefaultTenantId == id)
+                throw new ExceptionConflict("The default tenant cannot be deleted.");
+
             var entity = await GetEntityAsync(id);
             _dbContext.AppTenants.Remove(entity);
+            await _dbContext.SaveChangesAsync();
         }
 
         public override async Task<AppTenantEntity> GetEntityAsync(int id)
